refactor(ai): extract discovery/profiling internals to non-action module
SECURITY: Moved private helpers out of 'use server' file to prevent
unguarded functions from becoming publicly callable server actions.
